+++ common/src/main/java/com/yanny/ali/manager/AliServerRegistry.java
@@ -59,6 +59,13 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
     private final List<ILootModifier<?>> lootModifierMap = new LinkedList<>();
     private final ICommonUtils utils;
 
+    private final Set<Class<?>> missingFunctionTooltips = new HashSet<>();
+    private final Set<Class<?>> missingConditionTooltips = new HashSet<>();
+    private final Set<Class<?>> missingIngredientTooltips = new HashSet<>();
+    private final Set<Class<?>> missingItemSubPredicateTooltips = new HashSet<>();
+    private final Set<Class<?>> missingEntitySubPredicateTooltips = new HashSet<>();
+    private final Set<Class<?>> missingDataComponentTypeTooltips = new HashSet<>();
+
     private ServerLevel serverLevel;
     private LootContext lootContext;
 
@@ -204,7 +211,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         if (entryTooltipGetter != null) {
             return entryTooltipGetter.apply(utils, function);
         } else {
-            LOGGER.warn("Function tooltip for {} was not registered", function.getClass().getCanonicalName());
+            missingFunctionTooltips.add(function.getClass());
 
             try {
                 return GenericTooltipUtils.getMissingFunction(utils, function);
@@ -222,7 +229,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         if (entryTooltipGetter != null) {
             return entryTooltipGetter.apply(utils, condition);
         } else {
-            LOGGER.warn("Condition tooltip for {} was not registered", condition.getClass().getCanonicalName());
+            missingConditionTooltips.add(condition.getClass());
 
             try {
                 return GenericTooltipUtils.getMissingCondition(utils, condition);
@@ -240,7 +247,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         if (ingredientTooltipGetter != null) {
             return ingredientTooltipGetter.apply(utils, ingredient);
         } else {
-            LOGGER.warn("Ingredient tooltip for {} was not registered", ingredient.getClass().getCanonicalName());
+            missingIngredientTooltips.add(ingredient.getClass());
             return GenericTooltipUtils.getStringTooltip(utils, "ali.util.advanced_loot_info.missing", ingredient.getClass().getSimpleName());
         }
     }
@@ -252,7 +259,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         if (itemSubPredicateTooltipGetter != null) {
             return itemSubPredicateTooltipGetter.apply(utils, predicate);
         } else {
-            LOGGER.warn("Item sub predicate tooltip for {} was not registered", predicate.getClass().getCanonicalName());
+            missingItemSubPredicateTooltips.add(predicate.getClass());
             return GenericTooltipUtils.getStringTooltip(utils, "ali.util.advanced_loot_info.missing", predicate.getClass().getSimpleName());
         }
     }
@@ -264,7 +271,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         if (entitySubPredicateTooltipGetter != null) {
             return entitySubPredicateTooltipGetter.apply(utils, predicate);
         } else {
-            LOGGER.warn("Entity sub predicate tooltip for {} was not registered", predicate.getClass().getCanonicalName());
+            missingEntitySubPredicateTooltips.add(predicate.getClass());
             return GenericTooltipUtils.getStringTooltip(utils, "ali.util.advanced_loot_info.missing", predicate.getClass().getSimpleName());
         }
     }
@@ -276,7 +283,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         if (dataComponentTypeTooltipGetter != null) {
             return dataComponentTypeTooltipGetter.apply(utils, value);
         } else {
-            LOGGER.warn("Data component type tooltip for {} was not registered", type.getClass().getCanonicalName());
+            missingDataComponentTypeTooltips.add(type.getClass());
             return GenericTooltipUtils.getStringTooltip(utils, "ali.util.advanced_loot_info.missing", type.getClass().getSimpleName());
         }
     }
@@ -362,7 +369,7 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         return utils.createEntities(type, level);
     }
 
-    public void printServerInfo() {
+    public void printRegistrationInfo() {
         LOGGER.info("Registered {} entry item collectors", entryItemCollectorMap.size());
         LOGGER.info("Registered {} function item collectors", functionItemCollectorMap.size());
         LOGGER.info("Registered {} number converters", numberConverterMap.size());
@@ -377,5 +384,15 @@ public class AliServerRegistry implements IServerRegistry, IServerUtils {
         LOGGER.info("Registered {} count modifiers", countModifierMap.size());
         LOGGER.info("Registered {} item stack modifiers", itemStackModifierMap.size());
         LOGGER.info("Registered {} loot modifiers", lootModifierMap.size());
+    }
+
+    public void printRuntimeInfo() {
+        missingFunctionTooltips.forEach((t) -> LOGGER.warn("Missing function tooltip for {}", t.getCanonicalName()));
+        missingConditionTooltips.forEach((t) -> LOGGER.warn("Missing condition tooltip for {}", t.getCanonicalName()));
+        missingIngredientTooltips.forEach((t) -> LOGGER.warn("Missing ingredient tooltip for {}", t.getCanonicalName()));
+        missingItemSubPredicateTooltips.forEach((t) -> LOGGER.warn("Missing item sub predicate tooltip for {}", t.getCanonicalName()));
+        missingEntitySubPredicateTooltips.forEach((t) -> LOGGER.warn("Missing entity sub predicate tooltip for {}", t.getCanonicalName()));
+        missingDataComponentTypeTooltips.forEach((t) -> LOGGER.warn("Missing data component type tooltip for {}", t.getCanonicalName()));
+        LOGGER.info("Prepared {} loot tables", lootTableMap.size());
     }
 }