+++ common/src/test/java/com/yanny/ali/test/TooltipTestSuite.java
@@ -71,6 +71,7 @@ import java.util.concurrent.ExecutionException;
         EntitySubPredicateTooltipTest.class,
         ItemSubPredicateTooltipTest.class,
         DataComponentTooltipTest.class,
+        IngredientTooltipTest.class,
         TooltipTest.class
 })
 public class TooltipTestSuite {