+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## []
+
+- Fixed Ingredient tooltip
+- Make uppercased chest name from loot table name if missing translation
+
 ## [1.2.3]
 
 - Fixed crash on failed mixin